+++ src/app/core/graph-options/graph-options.component.ts
@@ -17,6 +17,7 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 import { map, startWith } from 'rxjs/operators';
 import { MatAutocompleteTrigger } from '@angular/material/autocomplete';
 
+import { NONE_TYPE } from '@angular/compiler';
 
 @Component({
   selector: 'app-graph-options',
@@ -60,6 +61,12 @@ export class GraphOptionsComponent implements OnInit {
   //flags
   flaggedData = [];
   showFlagOptions: Boolean = false;
+  showFlagOptionsX: Boolean = false;
+  showFlagOptionsY: Boolean = false;
+  submitAfterX: Boolean = false;
+  onlyYflags: Boolean = false;
+  differentYflags: Boolean = false;
+  maxFlagModalHeight;
   showUnflagOptions: Boolean = false;
   showLassoFlagOptions: Boolean = false;
   lasso: Boolean = false;
@@ -73,6 +80,8 @@ export class GraphOptionsComponent implements OnInit {
     xFlagControl: new FormControl(),
     yFlagControl: new FormControl(),
     xyFlagControl: new FormControl(),
+  });
+  public flagTypesX = new FormGroup({
     centralTendency: new FormControl(),
     outlier: new FormControl(),
     matrixInterference: new FormControl(),
@@ -80,17 +89,37 @@ export class GraphOptionsComponent implements OnInit {
     phytoChl: new FormControl(),
     unknown: new FormControl(),
   });
+  public flagTypesY = new FormGroup({
+    centralTendency: new FormControl(),
+    outlier: new FormControl(),
+    matrixInterference: new FormControl(),
+    dissolvedGTTotal: new FormControl(),
+    phytoChl: new FormControl(),
+    unknown: new FormControl(),
+  });
+  public sameXYFlag = new FormGroup({
+    diffY: new FormControl(),
+  });
   public flags$: Observable<any[]>;
   //Colors for all 4 flagging options
-  public unflaggedColor: string = 'rgb(242, 189, 161)';
+  public unflaggedColor: string = 'rgba(242, 204, 177, 0.2)';
   public xyFlaggedColor: string = 'rgb(0, 153, 0)';
   public xFlaggedColor: string = 'rgb(255, 0, 255)';
   public yFlaggedColor: string = 'rgb(0, 204, 204)';
+  public pointBorderColor: string = 'rgba(242, 204, 177, 1)';
   //Symbols for flagged vs unflagged
   public allColors = [];
   public allShapes = [];
-  public unflaggedSymbol: string = 'circle-open';
+  public allBorderWidths = [];
+  public allSizes = [];
+  public unflaggedSymbol: string = 'circle';
   public flaggedSymbol: string = 'circle';
+  public unflaggedSize: number = 16;
+  public flaggedSize: number = 12;
+  public flaggedBorderWidth: number = 0;
+  public unflaggedBorderWidth: number = 2;
+  public singlePointSelected: Boolean = true;
+
   public selectedPoints;
 
   //Intermediate data
@@ -157,7 +186,7 @@ export class GraphOptionsComponent implements OnInit {
   //Adjust the css (via resizeDivs()) when the window is resized
   @HostListener('window:resize')
   onResize() {
-    this.resizeDivs();
+    this.resizeDivs(true);
   }
 
   @HostListener("window:beforeunload")
@@ -173,7 +202,7 @@ export class GraphOptionsComponent implements OnInit {
 
   ngOnInit(): void {
     //Set the display according to the initial screen dimensions
-    this.resizeDivs();
+    this.resizeDivs(false);
     this.getDataForDropdowns();
     this.initiateGraphService();
     this.getUnits();
@@ -192,7 +221,7 @@ export class GraphOptionsComponent implements OnInit {
     this.componentDisplayService.usaBarCollapseSubject.subscribe(
       (usaBarBoolean) => {
         setTimeout(() => {
-          this.resizeDivs();
+          this.resizeDivs(true);
         }, 0.1);
       }
     );
@@ -247,6 +276,7 @@ export class GraphOptionsComponent implements OnInit {
     this.graphSelectionsService.flagsSubject.subscribe((flags) => {
       if (flags) {
         if (flags.length > 0) {
+          this.flaggedData = flags;
           this.disableEnable('flagBtn', true, true);
         } else {
           this.disableEnable('flagBtn', false, true);
@@ -408,32 +438,109 @@ export class GraphOptionsComponent implements OnInit {
   //Shrinks Graph Options panel into a button and expands the graph
   public collapseGraphOptions(collapsed: Boolean) {
     this.graphOptionsVisible = collapsed;
-    this.resizeDivs();
+    this.resizeDivs(true);
   }
 
   public createGraph(newPlot: Boolean) {
     if (newPlot) {
       this.allColors = this.graphSelectionsService.pointColors;
       this.allShapes = this.graphSelectionsService.pointSymbol;
+      this.allBorderWidths = this.graphSelectionsService.allBorderWidths;
+      this.allSizes = this.graphSelectionsService.allSizes;
     }
     //Designate div to put graph
     this.bivariatePlot = document.getElementById('graph');
 
-    var trace1 = {
+    var allData = {
       x: this.currentXaxisValues,
       y: this.currentYaxisValues,
       mode: 'markers',
       type: 'scatter',
-      //Keeping these here so that it's easy to add if we decide to implement in the future
-      //name: 'Sample 1',
-      //text: this.sid,
+      name: 'None',
+      showlegend: false,
+      hovertemplate: 'x: %{x} <br> y: %{y} <extra></extra>',
       textposition: 'bottom center',
-      marker: { size: 12, color: this.allColors, symbol: this.allShapes },
+      marker: {
+        size: this.allSizes,
+        color: this.allColors,
+        symbol: this.allShapes,
+        line: {
+          color: this.pointBorderColor,
+          width: this.allBorderWidths,
+        },
+      },
     };
 
-    var data = [trace1];
+    let AxisLegendFillerAllData = {
+      x: [null],
+      y: [null],
+      showlegend: true,
+      name: 'None',
+      mode: 'markers',
+      type: 'scatter',
+      marker: {
+        size: 16,
+        color: this.unflaggedColor,
+        symbol: this.unflaggedSymbol,
+        line: {
+          color: this.pointBorderColor,
+          width: this.allBorderWidths,
+        },
+      },
+    };
+
+    let AxisLegendFillerX = {
+      x: [null],
+      y: [null],
+      showlegend: true,
+      name: 'X-Axis',
+      mode: 'markers',
+      type: 'scatter',
+      marker: {
+        size: 12,
+        color: this.xFlaggedColor,
+        symbol: this.flaggedSymbol,
+      },
+    };
+
+    let AxisLegendFillerY = {
+      x: [null],
+      y: [null],
+      showlegend: true,
+      name: 'Y-Axis',
+      mode: 'markers',
+      type: 'scatter',
+      marker: {
+        size: 12,
+        color: this.yFlaggedColor,
+        symbol: this.flaggedSymbol,
+      },
+    };
+
+    let AxisLegendFillerXY = {
+      x: [null],
+      y: [null],
+      showlegend: true,
+      name: 'Both Axes',
+      mode: 'markers',
+      type: 'scatter',
+      marker: {
+        size: 12,
+        color: this.xyFlaggedColor,
+        symbol: this.flaggedSymbol,
+      },
+    };
+
+    var data = [
+      allData,
+      AxisLegendFillerAllData,
+      AxisLegendFillerX,
+      AxisLegendFillerY,
+      AxisLegendFillerXY,
+    ];
 
     var layout = {
+      dragmode: 'lasso',
       font: {
         size: 18,
       },
@@ -453,8 +560,12 @@ export class GraphOptionsComponent implements OnInit {
       },
       paper_bgcolor: 'rgba(255, 255, 255, 0)',
       plot_bgcolor: 'rgba(255, 255, 255, 0)',
-      showlegend: false,
-      legend: { bgcolor: 'rgba(255, 255, 255, 0)' },
+      showlegend: true,
+      legend: {
+        bgcolor: 'rgba(255, 255, 255, 0)',
+        title: { text: 'Flagged Data' },
+        font: { size: 14 },
+      },
       modebar: { bgcolor: 'rgba(255, 255, 255, 0)' },
       height: this.graphHeight,
       width: this.graphWidth,
@@ -489,14 +600,14 @@ export class GraphOptionsComponent implements OnInit {
       modeBarButtonsToRemove: ['select', 'resetscale', 'zoomin', 'zoomout'],
       modeBarButtonsToAdd: [
         {
-          name: 'Flag plotted data',
+          name: 'Flag all',
           icon: flagIcon,
           click: (e) => {
             this.flagAllData();
           },
         },
         {
-          name: 'Remove all flags',
+          name: 'Remove plotted flags',
           icon: removeIcon,
           click: (e) => {
             this.unflagAllData();
@@ -524,28 +635,88 @@ export class GraphOptionsComponent implements OnInit {
   unflagAllData() {
     this.unflagAll = true;
     this.showUnflagOptions = true;
+    this.disableEnableGraph(true);
     this.selectPoints();
   }
 
+  diffYflags(event) {
+    if (event.checked) {
+      this.differentYflags = true;
+    }
+    if (!event.checked) {
+      this.differentYflags = false;
+    }
+  }
+
+  public stringArray(flagString: String) {
+    let flagArray: String[] = null;
+    flagArray = flagString.split('; ');
+    return flagArray;
+  }
+
   closeFlagOptions() {
-    //Close flag options modal
-    this.showFlagOptions = false;
+    //Unfreeze graph and sidebar
     this.disableEnableGraph(false);
+    this.disableEnable('graph', true, false);
+    this.disableEnable('graphOptionsBackgroundID', true, false);
+
+    //Hide all flag option modals
+    this.showFlagOptions = false;
+    this.showFlagOptionsX = false;
+    this.showFlagOptionsY = false;
     this.showUnflagOptions = false;
+    this.onlyYflags = false;
+
+    //If lasso was used, redraw graph to reset color display
     if (this.lasso) {
       this.createGraph(false);
       this.lasso = false;
     }
+
+    //Reset single flag designation
+    this.singlePointSelected = false;
+
+    //Reset boolean that controls flagging all data at once
     this.flagAll = false;
 
-    //enable all features that were disabled when modal was open
-    this.disableEnable('graph', true, false);
-    this.disableEnable('graphOptionsBackgroundID', true, false);
-
-    //Reset form
+    //Reset x-axis/y-axis selection form
     this.axisFlagForm.get('xFlagControl').setValue(null);
     this.axisFlagForm.get('yFlagControl').setValue(null);
     this.axisFlagForm.get('xyFlagControl').setValue(null);
+
+    //Reset flag types x selection form
+    this.flagTypesX.get('centralTendency').setValue(null);
+    this.flagTypesX.get('outlier').setValue(null);
+    this.flagTypesX.get('matrixInterference').setValue(null);
+    this.flagTypesX.get('dissolvedGTTotal').setValue(null);
+    this.flagTypesX.get('phytoChl').setValue(null);
+    this.flagTypesX.get('unknown').setValue(null);
+
+    //Reset flag types y selection form
+    this.flagTypesY.get('centralTendency').setValue(null);
+    this.flagTypesY.get('outlier').setValue(null);
+    this.flagTypesY.get('matrixInterference').setValue(null);
+    this.flagTypesY.get('dissolvedGTTotal').setValue(null);
+    this.flagTypesY.get('phytoChl').setValue(null);
+    this.flagTypesY.get('unknown').setValue(null);
+
+    //Reset option to choose different y-axis flag types
+    this.sameXYFlag.get('diffY').setValue(null);
+    this.differentYflags = false;
+
+    //Reset the user input for x and y flag annotations
+    let inputX = document.getElementById(
+      'flagAnnotationX'
+    ) as HTMLInputElement | null;
+    if (inputX.value) {
+      inputX.value = '';
+    }
+    let inputY = document.getElementById(
+      'flagAnnotationY'
+    ) as HTMLInputElement | null;
+    if (inputY.value) {
+      inputY.value = '';
+    }
   }
 
   //Disables or enables clickable features
@@ -572,7 +743,6 @@ export class GraphOptionsComponent implements OnInit {
     const plotlyjs = document.getElementsByClassName('js-plotly-plot');
     for (let el in plotlyjs) {
       if (plotlyjs[el]['style'] !== undefined && enable) {
-        console.log('enable true');
         plotlyjs[el]['style'].pointerEvents = 'unset';
       } else if (plotlyjs[el]['style'] !== undefined && !enable) {
         plotlyjs[el]['style'].pointerEvents = 'none';
@@ -595,12 +765,18 @@ export class GraphOptionsComponent implements OnInit {
     color: String,
     axis: String,
     symbol: String,
-    flagTypes,
-    annotation: String
+    border: Number,
+    size: Number,
+    flagTypesX,
+    flagTypesY,
+    annotationX: String,
+    annotationY: String
   ) {
     let updateGraphCalled = true;
     let colors = this.allColors;
     let symbols = this.allShapes;
+    let borders = this.allBorderWidths;
+    let sizes = this.allSizes;
     let numPts;
 
     //If flagAll button is clicked, need to get attributes outside of Plotly
@@ -642,6 +818,8 @@ export class GraphOptionsComponent implements OnInit {
       colors[pointIndex] = color;
       //Change the symbol of the point at the correct index (flagged pts become filled circles; unflagged becomes hollow circle)
       symbols[pointIndex] = symbol;
+      borders[pointIndex] = border;
+      sizes[pointIndex] = size;
 
       //if only the x-axis is selected, make sure the y-value at that point isn't in the flaggedData array
       //if neither are selected, ensure that neither are in the flaggedData array
@@ -688,8 +866,8 @@ export class GraphOptionsComponent implements OnInit {
         this.graphSelectionsService.allGraphDataXSubject.subscribe((data) => {
           if (updateGraphCalled) {
             tempData = data;
-            tempData[pointIndex]['flagType'] = flagTypes;
-            tempData[pointIndex]['annotation'] = annotation;
+            tempData[pointIndex]['flagType'] = flagTypesX;
+            tempData[pointIndex]['annotation'] = annotationX;
             if (!existingDupX) {
               this.flaggedData.push(tempData[pointIndex]);
               this.graphSelectionsService.flagsSubject.next(this.flaggedData);
@@ -702,8 +880,8 @@ export class GraphOptionsComponent implements OnInit {
         this.graphSelectionsService.allGraphDataYSubject.subscribe((data) => {
           if (updateGraphCalled) {
             tempData = data;
-            tempData[pointIndex]['flagType'] = flagTypes;
-            tempData[pointIndex]['annotation'] = annotation;
+            tempData[pointIndex]['flagType'] = flagTypesY;
+            tempData[pointIndex]['annotation'] = annotationY;
             if (!existingDupY) {
               this.flaggedData.push(tempData[pointIndex]);
               this.graphSelectionsService.flagsSubject.next(this.flaggedData);
@@ -713,21 +891,28 @@ export class GraphOptionsComponent implements OnInit {
       }
     }
 
-    //New styling for new plot
-    let update = {
-      marker: { color: colors, size: 12, symbol: symbols },
-    };
-
     this.allColors = colors;
     this.allShapes = symbols;
+    this.allBorderWidths = borders;
+    this.allSizes = sizes;
 
-    //Change the color on the graph
-    if (!this.lasso) {
-      Plotly.restyle('graph', update);
-    }
-    //Override the default Plotly post-lasso view by re-drawing graph
-    if (this.lasso) {
+    //If only one point was selected, do not need to re-create graph
+    if (!this.singlePointSelected) {
       this.createGraph(false);
+    } else {
+      //New styling for new plot
+      let update = {
+        marker: {
+          size: this.allSizes,
+          color: this.allColors,
+          symbol: this.allShapes,
+          line: {
+            color: this.pointBorderColor,
+            width: this.allBorderWidths,
+          },
+        },
+      };
+      Plotly.restyle('graph', update, [1]);
     }
 
     // storing the cyanFlags in browser local storage
@@ -745,21 +930,114 @@ export class GraphOptionsComponent implements OnInit {
   }
 
   sumbitUnflagSelections() {
-    this.updateGraph(this.unflaggedColor, 'none', this.unflaggedSymbol, '', '');
+    this.updateGraph(
+      this.unflaggedColor,
+      'none',
+      this.unflaggedSymbol,
+      this.unflaggedBorderWidth,
+      this.unflaggedSize,
+      ', ',
+      ',',
+      ',',
+      ','
+    );
     this.showUnflagOptions = false;
     this.closeFlagOptions();
+  }
+
+  //Disable/enable 'Next' button depending on whether or not an axis is checked
+  public xyFlagClicked() {
+    let xChecked = this.axisFlagForm.get('xFlagControl').value;
+    let yChecked = this.axisFlagForm.get('yFlagControl').value;
+    let xyChecked = this.axisFlagForm.get('xyFlagControl').value;
+
+    if (xChecked || yChecked || xyChecked) {
+      this.disableEnable('continueToFlagOptions', true, true);
+    } else {
+      this.disableEnable('continueToFlagOptions', false, true);
+    }
+  }
+
+  public goToFlagTypes() {
+    let xChecked = this.axisFlagForm.get('xFlagControl').value;
+    let yChecked = this.axisFlagForm.get('yFlagControl').value;
+    let xyChecked = this.axisFlagForm.get('xyFlagControl').value;
+    this.showFlagOptions = false;
+
+    if (xChecked || xyChecked) {
+      this.showFlagOptionsX = true;
+    }
+    if (!xChecked && !xyChecked && yChecked) {
+      this.onlyYflags = true;
+      this.showFlagOptionsY = true;
+    }
+    if (yChecked) {
+      this.submitAfterX = false;
+    }
+    if (!yChecked) {
+      this.submitAfterX = true;
+    }
+
+    setTimeout(() => {
+      document.getElementById('flagModals').style.height = 'auto';
+      this.maxFlagModalHeight =
+        document.getElementById('flagModals').clientHeight - 20;
+      this.resizeDivs(false);
+    }, 1);
+  }
+
+  public goToFlagTypesY() {
+    this.showFlagOptionsX = false;
+    this.showFlagOptionsY = true;
+
+    this.maxFlagModalHeight =
+      document.getElementById('flagModals').clientHeight;
+
+    setTimeout(() => {
+      document.getElementById('flagModals').style.height = 'auto';
+      this.maxFlagModalHeight =
+        document.getElementById('flagModals').clientHeight - 20;
+      this.resizeDivs(true);
+    }, 1);
+  }
+
+  public getAnnotation(axis: String) {
+    let annotation = '';
+    let input;
+    if (axis == 'x') {
+      input = document.getElementById(
+        'flagAnnotationX'
+      ) as HTMLInputElement | null;
+    }
+    if (axis == 'y') {
+      input = document.getElementById(
+        'flagAnnotationY'
+      ) as HTMLInputElement | null;
+    }
+    annotation = input?.value;
+    annotation = annotation.replace(/,/g, ';');
+
+    return annotation;
   }
 
   //Triggered when the 'Submit' button is clicked in the flag modal
   submitFlagSelections() {
     //Capture user input in the annotation box
-    let annotation = '';
-    let input = document.getElementById(
-      'flagAnnotation'
-    ) as HTMLInputElement | null;
-    annotation = input?.value;
-    annotation = annotation.replace(/,/g, ';');
-    let flagTypes = this.flagTypes();
+    let annotationX = this.getAnnotation('x');
+    let annotationY;
+    let flagTypesX = this.flagTypes('x');
+    let flagTypesY;
+    //if user chose different flag types for the y-axis, get responses from form
+    if (this.differentYflags || this.onlyYflags) {
+      flagTypesY = this.flagTypes('y');
+      annotationY = this.getAnnotation('y');
+    }
+    this.stringArray(flagTypesX);
+    //if user chose to use the same responses for both axes, duplicate x responses
+    if (!this.differentYflags && !this.onlyYflags) {
+      flagTypesY = flagTypesX;
+      annotationY = annotationX;
+    }
     if (!this.sameQuery) {
       let xChecked = this.axisFlagForm.get('xFlagControl').value;
       let yChecked = this.axisFlagForm.get('yFlagControl').value;
@@ -769,8 +1047,12 @@ export class GraphOptionsComponent implements OnInit {
           this.xyFlaggedColor,
           'both',
           this.flaggedSymbol,
-          flagTypes,
-          annotation
+          this.flaggedBorderWidth,
+          this.flaggedSize,
+          flagTypesX,
+          flagTypesY,
+          annotationX,
+          annotationY
         );
       }
       //Y checked; x not checked
@@ -779,8 +1061,12 @@ export class GraphOptionsComponent implements OnInit {
           this.yFlaggedColor,
           'y',
           this.flaggedSymbol,
-          flagTypes,
-          annotation
+          this.flaggedBorderWidth,
+          this.flaggedSize,
+          flagTypesX,
+          flagTypesY,
+          annotationX,
+          annotationY
         );
       }
 
@@ -790,8 +1076,12 @@ export class GraphOptionsComponent implements OnInit {
           this.xFlaggedColor,
           'x',
           this.flaggedSymbol,
-          flagTypes,
-          annotation
+          this.flaggedBorderWidth,
+          this.flaggedSize,
+          flagTypesX,
+          flagTypesY,
+          annotationX,
+          annotationY
         );
       }
 
@@ -801,11 +1091,14 @@ export class GraphOptionsComponent implements OnInit {
           this.unflaggedColor,
           'none',
           this.unflaggedSymbol,
-          flagTypes,
-          annotation
+          this.unflaggedBorderWidth,
+          this.unflaggedSize,
+          flagTypesX,
+          flagTypesY,
+          annotationX,
+          annotationY
         );
       }
-      this.flagTypes();
     }
     if (this.sameQuery) {
       let xyChecked = this.axisFlagForm.get('xyFlagControl').value;
@@ -815,36 +1108,62 @@ export class GraphOptionsComponent implements OnInit {
           this.xyFlaggedColor,
           'x',
           this.flaggedSymbol,
-          flagTypes,
-          annotation
+          this.flaggedBorderWidth,
+          this.unflaggedSize,
+          flagTypesX,
+          flagTypesY,
+          annotationX,
+          annotationY
         );
       } else {
         this.updateGraph(
           this.unflaggedColor,
           'none',
           this.unflaggedSymbol,
-          flagTypes,
-          annotation
+          this.unflaggedBorderWidth,
+          this.unflaggedSize,
+          flagTypesX,
+          flagTypesY,
+          annotationX,
+          annotationY
         );
       }
     }
 
-    //Clear annotation form
-    if (input.value) {
-      input.value = null;
-    }
     //Close flag modal and clear selections
     this.closeFlagOptions();
   }
 
-  public flagTypes() {
-    let centralTendency = this.axisFlagForm.get('centralTendency').value;
-    let outlier = this.axisFlagForm.get('outlier').value;
-    let matrixInterference = this.axisFlagForm.get('matrixInterference').value;
-    let dissolvedGTTotal = this.axisFlagForm.get('dissolvedGTTotal').value;
-    let phytoChl = this.axisFlagForm.get('phytoChl').value;
-    let unknown = this.axisFlagForm.get('unknown').value;
+  public flagTypes(axis: String) {
+    //All flag type options
+    let centralTendency;
+    let outlier;
+    let matrixInterference;
+    let dissolvedGTTotal;
+    let phytoChl;
+    let unknown;
 
+    //Get x flag types
+    if (axis == 'x') {
+      centralTendency = this.flagTypesX.get('centralTendency').value;
+      outlier = this.flagTypesX.get('outlier').value;
+      matrixInterference = this.flagTypesX.get('matrixInterference').value;
+      dissolvedGTTotal = this.flagTypesX.get('dissolvedGTTotal').value;
+      phytoChl = this.flagTypesX.get('phytoChl').value;
+      unknown = this.flagTypesX.get('unknown').value;
+    }
+
+    //Get y flag types
+    if (axis == 'y') {
+      centralTendency = this.flagTypesY.get('centralTendency').value;
+      outlier = this.flagTypesY.get('outlier').value;
+      matrixInterference = this.flagTypesY.get('matrixInterference').value;
+      dissolvedGTTotal = this.flagTypesY.get('dissolvedGTTotal').value;
+      phytoChl = this.flagTypesY.get('phytoChl').value;
+      unknown = this.flagTypesY.get('unknown').value;
+    }
+
+    //Convert flag types selections into a string for the csv
     let flagTypes = '';
     if (centralTendency) {
       flagTypes += 'Central tendency; ';
@@ -862,19 +1181,19 @@ export class GraphOptionsComponent implements OnInit {
       flagTypes += 'Phytoplankton vs Chl; ';
     }
     if (unknown) {
-      flagTypes += 'Unknown';
+      flagTypes += 'Unknown; ';
     }
-    this.axisFlagForm.get('centralTendency').setValue(null);
-    this.axisFlagForm.get('outlier').setValue(null);
-    this.axisFlagForm.get('matrixInterference').setValue(null);
-    this.axisFlagForm.get('dissolvedGTTotal').setValue(null);
-    this.axisFlagForm.get('phytoChl').setValue(null);
-    this.axisFlagForm.get('unknown').setValue(null);
+
+    //remove final semicolon and space
+    flagTypes = flagTypes.slice(0, -2);
+
+    //Return string of all selected flag types for one axis
     return flagTypes;
   }
 
   public initiateSelectPoints() {
     this.bivariatePlot.on('plotly_click', (selectedPoints) => {
+      this.singlePointSelected = true;
       this.selectPoints();
       //If there is a flag at the selected point, pre-check the boxes in the flag options modal
       this.axisFlagFormCheckBoxes(selectedPoints.points);
@@ -884,20 +1203,29 @@ export class GraphOptionsComponent implements OnInit {
       this.disableEnableGraph(true);
     });
     this.bivariatePlot.on('plotly_selected', (selectedPoints) => {
-      this.selectedPoints = selectedPoints.points;
-      //Open flag options modal
-      this.showFlagOptions = true;
-      this.lasso = true;
-      this.selectPoints();
-      this.disableEnableGraph(true);
+      if (selectedPoints) {
+        this.selectedPoints = selectedPoints.points;
+        //Open flag options modal
+        this.showFlagOptions = true;
+        this.lasso = true;
+        this.selectPoints();
+        this.disableEnableGraph(true);
+      }
     });
   }
 
   public selectPoints() {
+    document.getElementById('flagModals').style.height = 'auto';
     //Prevent user from clicking on features outside the modal
     this.disableEnable('graph', false, false);
     this.disableEnable('graphOptionsBackgroundID', false, false);
-    //Plotly.react(this.bivariatePlot, {}, {'staticPlot': 'true'});
+
+    //Get original flag modal height
+    setTimeout(() => {
+      this.maxFlagModalHeight =
+        document.getElementById('flagModals').clientHeight - 20;
+      this.resizeDivs(false);
+    }, 1);
   }
 
   //If there is a flag at the selected point, pre-check the boxes in the flag options modal
@@ -908,23 +1236,110 @@ export class GraphOptionsComponent implements OnInit {
       this.axisFlagForm.get('xFlagControl').setValue(true);
       this.axisFlagForm.get('yFlagControl').setValue(true);
       this.axisFlagForm.get('xyFlagControl').setValue(true);
+      this.autoCheckFlagTypes('x', selectedPoints);
+      this.autoCheckFlagTypes('y', selectedPoints);
     }
     if (selectedColor == this.xFlaggedColor) {
       //check the x box
       this.axisFlagForm.get('xFlagControl').setValue(true);
+      this.autoCheckFlagTypes('x', selectedPoints);
     }
     if (selectedColor == this.yFlaggedColor) {
       //check the y box
       this.axisFlagForm.get('yFlagControl').setValue(true);
+      this.autoCheckFlagTypes('y', selectedPoints);
+    }
+    //Disable/enable 'Next' button depending on whether or not an axis is checked
+    this.xyFlagClicked();
+  }
+
+  autoCheckFlagTypes(axis: String, selectedPoints) {
+    if (axis === 'x') {
+      this.graphSelectionsService.allGraphDataXSubject.subscribe((xdata) => {
+        let selectedRcodeX = xdata[selectedPoints[0].pointIndex].rcode;
+        for (let i = 0; i < this.flaggedData.length; i++) {
+          if (this.flaggedData[i].rcode == selectedRcodeX) {
+            let currentAnnotation = this.flaggedData[i].annotation;
+            let insertAnnotation = document.getElementById(
+              'flagAnnotationX'
+            ) as HTMLInputElement | null;
+            insertAnnotation.value = currentAnnotation;
+            let currentFlagType = this.flaggedData[i].flagType;
+            let flagArray = this.stringArray(currentFlagType);
+            if (flagArray) {
+              for (let j = 0; j < flagArray.length; j++) {
+                if (flagArray[j] == 'Central tendency') {
+                  this.flagTypesX.get('centralTendency').setValue(true);
+                }
+                if (flagArray[j] == 'Outlier') {
+                  this.flagTypesX.get('outlier').setValue(true);
+                }
+                if (flagArray[j] == 'Matrix or recovery problem') {
+                  this.flagTypesX.get('matrixInterference').setValue(true);
+                }
+                if (flagArray[j] == 'Dissolved result > Total') {
+                  this.flagTypesX.get('dissolvedGTTotal').setValue(true);
+                }
+                if (flagArray[j] == 'Phytoplankton vs Chl') {
+                  this.flagTypesX.get('phytoChl').setValue(true);
+                }
+                if (flagArray[j] == 'Unknown') {
+                  this.flagTypesX.get('unknown').setValue(true);
+                }
+              }
+            }
+          }
+        }
+      });
+    }
+    if (axis === 'y') {
+      this.graphSelectionsService.allGraphDataYSubject.subscribe((ydata) => {
+        let selectedRcodeY = ydata[selectedPoints[0].pointIndex].rcode;
+        for (let i = 0; i < this.flaggedData.length; i++) {
+          if (this.flaggedData[i].rcode == selectedRcodeY) {
+            let currentAnnotation = this.flaggedData[i].annotation;
+            let insertAnnotation = document.getElementById(
+              'flagAnnotationY'
+            ) as HTMLInputElement | null;
+            insertAnnotation.value = currentAnnotation;
+            let currentFlagType = this.flaggedData[i].flagType;
+            let flagArray = this.stringArray(currentFlagType);
+            if (flagArray) {
+              for (let j = 0; j < flagArray.length; j++) {
+                if (flagArray[j] == 'Central tendency') {
+                  this.flagTypesY.get('centralTendency').setValue(true);
+                }
+                if (flagArray[j] == 'Outlier') {
+                  this.flagTypesY.get('outlier').setValue(true);
+                }
+                if (flagArray[j] == 'Matrix or recovery problem') {
+                  this.flagTypesY.get('matrixInterference').setValue(true);
+                }
+                if (flagArray[j] == 'Dissolved result > Total') {
+                  this.flagTypesY.get('dissolvedGTTotal').setValue(true);
+                }
+                if (flagArray[j] == 'Phytoplankton vs Chl') {
+                  this.flagTypesY.get('phytoChl').setValue(true);
+                }
+                if (flagArray[j] == 'Unknown') {
+                  this.flagTypesY.get('unknown').setValue(true);
+                }
+              }
+            }
+          }
+        }
+      });
     }
   }
 
   //Called when user clicks 'Plot Data'
   public clickPlotData() {
+    //Whenever user attempts to create a new graph, clear previous one even if it's a failed query
+    this.showGraph = false;
     this.alreadyGraphed = false;
     //Get parameter and method user selections
-    let tempP_X = this.graphSelectionsForm.get('ParametersX').value.pcode;
-    let tempP_Y = this.graphSelectionsForm.get('ParametersY').value.pcode;
+    let tempP_X_value = this.graphSelectionsForm.get('ParametersX').value;
+    let tempP_Y_value = this.graphSelectionsForm.get('ParametersY').value;
     let tempM_X = this.graphSelectionsForm.get('MethodsX').value;
     let tempM_Y = this.graphSelectionsForm.get('MethodsY').value;
 
@@ -939,8 +1354,8 @@ export class GraphOptionsComponent implements OnInit {
 
     //If any parameter or method is left blank, prompt user to make a selection
     if (
-      tempP_X === null ||
-      tempP_Y === null ||
+      tempP_X_value === null ||
+      tempP_Y_value === null ||
       tempM_X == null ||
       tempM_Y == null
     ) {
@@ -960,7 +1375,7 @@ export class GraphOptionsComponent implements OnInit {
       this.disableEnable('graphDataDownloadBtn', false, true);
       this.showGraph = false;
       this.populateGraphData();
-      this.resizeDivs();
+      this.resizeDivs(true);
 
       //minimize graph options panel if the screen width is small
       let windowWidth = window.innerWidth;
@@ -1222,7 +1637,7 @@ export class GraphOptionsComponent implements OnInit {
   }
 
   //Every time the window is resized, change size and position of elements accordingly
-  public resizeDivs() {
+  public resizeDivs(redrawGraph) {
     //get map height
     let mapContainer = document.getElementById('mapContainer');
     let mapHeight = parseInt(window.getComputedStyle(mapContainer).height);
@@ -1238,8 +1653,17 @@ export class GraphOptionsComponent implements OnInit {
       'graphOptionsCollapsedID'
     );
 
-    this.graphHeight = 0.7 * mapHeight;
-    graphBackgroundID.style.height = (0.7 * mapHeight).toString() + 'px';
+    let flagModals = document.getElementById('flagModals');
+
+    if (this.maxFlagModalHeight > mapHeight - 130) {
+      let newFlagModalHeight = (mapHeight - 130).toString() + 'px';
+      flagModals.style.height = newFlagModalHeight;
+    } else {
+      flagModals.style.height = 'auto';
+    }
+
+    this.graphHeight = 0.99 * mapHeight - 100;
+    graphBackgroundID.style.height = (0.99 * mapHeight - 100).toString() + 'px';
 
     if (windowWidth < 900) {
       graphOptionsBackgroundID.classList.remove('marginLeftFullWidth');
@@ -1309,7 +1733,7 @@ export class GraphOptionsComponent implements OnInit {
       //this.graphMargins = 20;
     }
 
-    if (this.showGraph) {
+    if (this.showGraph && redrawGraph) {
       this.createGraph(false);
     }
   }